Initial commit for marshalling 'entities' into dynamo
The idea is to take any struct or map a user wants and simply allow it
to be stored in dynamo, without the use of the 'Attribute' type. This
method is simpler than MarshalAttributes, and also allows for nested
data types.

This first commit is a work in progress, and simply serializes the
scalar types.