Implement RetryPolicy
A RetryPolicy encapsulates a strategy for implementing client retries.
The desire is to apply the same default behavior as the AWS SDKs. This
functionality is being added first for use by DynamoDB, but with the
hope that it becomes standard across services.

cf: https://github.com/aws/aws-sdk-java/blob/master/aws-java-sdk-core/src/main/java/com/amazonaws/retry/PredefinedRetryPolicies.java

The randomness aspect still needs to be applied to the delay time
calculation for the default policy.