Fix issue where auth token can become empty
If we see a token is near expiry, we call GetAuth to refresh it, and
blindly overwrite our token without checking for error. If an error
is returned by GetAuth, we're left with the zero value Auth object,
which is empty. This means all future service calls will fail, and
since the token is now the empty string, we won't try to call GetAuth
again.

This change makes it so we only store the result of GetAuth if it
returns successfully. If it fails, we will keep using our existing
credentials, but each subsequent call to Token will try again to fetch
new credentials.