Add UpdateExpressionUpdateAttributes to support full UpdateExpression grammar
The existing UpdateAttributes method on Table could potentially be modified
to support a limited subset of UpdateExpressions, but it can't currently
support semantics like "SET #key1=:val1, DELETE #key2=:val2" that are permitted
by the DynamoDB API.

Providing a method that allows us to pass an UpdateExpression struct will support
arbitrary UpdateExpressions at the cost of the complexity to actually write one.
The added complexity only appears when we actually need it, so this seems like a
good trade off. Examples are provided in the associated test code.